Refactor login
Separate functionality for social (Google) and password login
